Add files via upload
A myocardial infarction (also called a heart attack) is a dangerous condition caused by a lack of blood flow to your heart muscle.
The mifem data frame has 1295 rows and 10 columns. This is a Dataset of females having coronary heart disease (CHD). you have to predict with the given information whether the female is dead or alive so as to discover important factors that should be considered crucial in the treatment of the disease. We use CART and RF and compare the models' performances in train and test sets.